Split ast.kt
