+++ src/main/kotlin/org/j2c/ast/NAThrow.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NAThrow(val v: Node): Node("NAThrow") {
+    override fun toString() = "throw $v"
+}
