+++ src/main/kotlin/org/j2c/ast/NArrayLength.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NArrayLength(val array: Node): Node("NArrayLength") {
+    override fun toString() = "#$array"
+}
