+++ src/main/kotlin/org/j2c/ast/NAssignment.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NAssignment(val dest: String, val v: Node): Node("NAssignment") {
+    override fun toString() = "$dest = $v"
+}
