+++ src/main/kotlin/org/j2c/ast/NBinOp.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NBinOp(val type: String, val op: String, val left: Node, val right: Node): Node("NBinOp") {
+    override fun toString() = "$left $op $right"
+}
