+++ src/main/kotlin/org/j2c/ast/NBoundReference.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NBoundReference(val obj: Node, val field: String): Node("NBoundReference") {
+    override fun toString() = "$obj.$field"
+}
