+++ src/main/kotlin/org/j2c/ast/NNew.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NNew(val clazz: String): Node("NNew") {
+    override fun toString() = "new $clazz"
+}
