+++ src/main/kotlin/org/j2c/ast/NNull.kt
@@ -1,0 +1,6 @@
+package org.j2c.ast
+
+// Primitive types put in Java to distinguish primitives from objects
+class NNull: Node("NNull") {
+    override fun toString() = "null"
+}
