+++ src/main/kotlin/org/j2c/ast/NReference.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NReference(val identifier: String): Node("NReference") {
+    override fun toString() = identifier
+}
