+++ src/main/kotlin/org/j2c/ast/NReturn.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NReturn: Node("NReturn") {
+    override fun toString() = "return"
+}
