+++ src/main/kotlin/org/j2c/ast/NStaticAssignment.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NStaticAssignment(val field: String, val v: Node): Node("NStaticAssignment") {
+    override fun toString() = "$field = $v"
+}
