+++ src/main/kotlin/org/j2c/ast/NStaticReference.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NStaticReference(val field: String): Node("NStaticReference") {
+    override fun toString() = field
+}
