+++ src/main/kotlin/org/j2c/ast/NValueReturn.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+class NValueReturn(val type: String, val v: Node): Node("NValueReturn") {
+    override fun toString() = "return $v"
+}
