+++ src/main/kotlin/org/j2c/ast/Node.kt
@@ -1,0 +1,5 @@
+package org.j2c.ast
+
+abstract class Node(val astName: String) {
+    abstract override fun toString(): String
+}
